consolidated StudentDashboard and AdminDashboard into Dashboard.js with isAdmin prop passed in from App.js as conditional for three differences in the file
